--Empty Message--

+++ blmarket/RockPaperScissors.cpp
@@ -79,7 +79,7 @@ public:
 
         N = 0;
         for(int i=0;i<size(rockProb);i++) {
-            addone((double)rockProb[i]/300.0, paperProb[i] / 300.0, scissorsProb[i] / 300.0);
+            addone((double)rockProb[i]/300.0, (double)paperProb[i] / 300.0, (double)scissorsProb[i] / 300.0);
         }
         for(int i=0;i<55;i++) for(int j=0;j<55;j++) for(int k=0;k<55;k++) nex[i][j][k] = -1.0;
